+++ idb_shim/lib/idb_shim.dart
@@ -8,4 +8,4 @@ library idb_shim;
 
 export 'idb.dart';
 export 'idb_client_memory.dart' show idbFactoryMemory, idbFactoryMemoryFs;
-export 'src/native/idb_native.dart' show idbFactoryNative;
+export 'src/native_web/idb_native.dart' show idbFactoryNative;