+++ idb_shim/lib/src/native_web/idb_native.dart
@@ -1,1 +1,2 @@
-export 'idb_native_stub.dart' if (dart.library.html) 'idb_native_web.dart';
+export 'idb_native_stub.dart'
+    if (dart.library.js_interop) 'idb_native_web.dart';