+++ idb_shim/lib/src/sembast/sembast_database.dart
@@ -129,14 +129,14 @@ class DatabaseSembast extends IdbDatabaseBase with DatabaseWithMetaMixin {
 
     // devPrint('open ${onUpgradeNeeded} ${onUpgradeNeeded != null ? 'NOT NULL': 'NULL'}');
     if (sembastDebug) {
-      print(
-          'open2 $onUpgradeNeeded ${onUpgradeNeeded != null ? 'NOT NULL' : 'NULL'}');
+      idbLog(
+          'open $onUpgradeNeeded ${onUpgradeNeeded != null ? 'NOT NULL' : 'NULL'}');
     }
     // Open the sembast database
     db = await sdbFactory.openDatabase(factory.getDbPath(name), version: 1,
         onVersionChanged: (db, oldVersion, newVersion) {
       if (sembastDebug) {
-        print('changing ${db.path} $oldVersion -> $newVersion');
+        idbLog('changing ${db.path} $oldVersion -> $newVersion');
       }
     });
     try {