+++ idb_test/test/web/test_runner_client_native_html_test.dart
@@ -1,8 +1,8 @@
 @TestOn('browser')
-library idb_shim.test_runner_client_native_test;
+library;
 
 import 'package:idb_shim/idb_client.dart';
-import 'package:idb_shim/idb_client_native.dart';
+import 'package:idb_shim/idb_client_native_html.dart';
 import 'package:idb_shim/utils/idb_utils.dart';
 import 'package:idb_test/idb_test_common.dart';
 import 'package:idb_test/test_runner.dart';